backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "labuan-fsa-backend"
version = "1.0.0"
description = "Labuan FSA E-Submission System Backend API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Labuan FSA Development Team"},
]
keywords = ["fastapi", "form-submission", "labuan-fsa", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "boto3>=1.29.7",  # AWS S3 integration
    "azure-storage-blob>=12.19.0",  # Azure Blob Storage integration
    "google-cloud-storage>=2.14.0",  # GCP Cloud Storage integration
    "sendgrid>=6.11.0",  # Email service
    "jinja2>=3.1.2",  # Email templates
    "toml>=0.10.2",  # TOML configuration
    "httpx>=0.25.2",  # HTTP client for external APIs
    "redis>=5.0.1",  # Caching and rate limiting
    "structlog>=23.2.0",  # Structured logging
    "prometheus-client>=0.19.0",  # Metrics
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.2",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/labuan-fsa/e-submission-system"
Documentation = "https://docs.labuanfsa.gov.my"
Repository = "https://github.com/labuan-fsa/e-submission-system"

[tool.setuptools]
packages = ["labuan_fsa"]

[tool.setuptools.package-data]
labuan_fsa = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--cov=labuan_fsa",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]

